Refactoring clients examples
