+++ examples/clients/README.md
@@ -1,0 +1,6 @@
+![SendGrid Logo](https://uiux.s3.amazonaws.com/2016-logos/email-logo%402x.png)
+
+This folder contains various examples on using the CLIENTS endpoint of SendGrid with Java:
+
+* [Retrieve email statistics by client type (GET /clients/stats)](GetEmailStatistics.java)
+* [Retrieve stats by a specific client type (GET /clients/{client_type}/stats)](GetEmailStatisticsByClientType.java)
